+++ CHANGELOG.rst
@@ -2,6 +2,13 @@ django-s3-storage changelog
 ===========================
 
 
+0.9.2
+-----
+
+- Added settings for fine-grained control over browser caching (@etianen).
+- Added settings for adding a prefix to all keys.
+
+
 0.9.1
 -----
 
@@ -12,4 +19,4 @@ django-s3-storage changelog
 0.9.0
 -----
 
-- First production release.
+- First production release (@etianen).